move jobs/commands to dedicated cron folder
