+++ packages/twenty-server/src/command/command.module.ts
@@ -1,21 +1,21 @@
 import { Module } from '@nestjs/common';
 
 import { DatabaseCommandModule } from 'src/database/commands/database-command.module';
-import { MessagingCommandModule } from 'src/modules/messaging/commands/messaging-command.module';
 import { WorkspaceHealthCommandModule } from 'src/engine/workspace-manager/workspace-health/commands/workspace-health-command.module';
 import { WorkspaceCleanerModule } from 'src/engine/workspace-manager/workspace-cleaner/workspace-cleaner.module';
-import { WorkspaceCalendarSyncCommandsModule } from 'src/modules/calendar/commands/workspace-calendar-sync-commands.module';
+import { CalendarCronCommandsModule } from 'src/modules/calendar/crons/commands/calendar-cron-commands.module';
 import { AppModule } from 'src/app.module';
 import { WorkspaceMigrationRunnerCommandsModule } from 'src/engine/workspace-manager/workspace-migration-runner/commands/workspace-sync-metadata-commands.module';
 import { WorkspaceSyncMetadataCommandsModule } from 'src/engine/workspace-manager/workspace-sync-metadata/commands/workspace-sync-metadata-commands.module';
+import { MessagingCronCommandsModule } from 'src/modules/messaging/crons/commands/messaging-cron-commands.module';
 
 @Module({
   imports: [
     AppModule,
     WorkspaceSyncMetadataCommandsModule,
     DatabaseCommandModule,
-    MessagingCommandModule,
-    WorkspaceCalendarSyncCommandsModule,
+    MessagingCronCommandsModule,
+    CalendarCronCommandsModule,
     WorkspaceCleanerModule,
     WorkspaceHealthCommandModule,
     WorkspaceMigrationRunnerCommandsModule,
