--- packages/twenty-server/src/modules/calendar/commands/crons/patterns/google-calendar-sync.cron.pattern.ts
@@ -1,1 +1,0 @@
-export const googleCalendarSyncCronPattern = '*/5 * * * *';
