--- packages/twenty-server/src/modules/messaging/commands/crons/patterns/gmail-partial-sync.cron.pattern.ts
@@ -1,1 +1,0 @@
-export const gmailPartialSyncCronPattern = '*/5 * * * *';
