+++ packages/twenty-server/src/modules/messaging/crons/commands/gmail-partial-sync.cron.command.ts
@@ -4,13 +4,16 @@ import { Command, CommandRunner } from 'nest-commander';
 
 import { MessageQueue } from 'src/engine/integrations/message-queue/message-queue.constants';
 import { MessageQueueService } from 'src/engine/integrations/message-queue/services/message-queue.service';
-import { GmailFetchMessagesFromCacheCronJob } from 'src/modules/messaging/jobs/crons/gmail-fetch-messages-from-cache.cron.job';
+import { GmailPartialSyncCronJob } from 'src/modules/messaging/crons/jobs/gmail-partial-sync.cron.job';
+
+const GMAIL_PARTIAL_SYNC_CRON_PATTERN = '*/5 * * * *';
 
 @Command({
-  name: 'cron:messaging:gmail-fetch-messages-from-cache',
-  description: 'Starts a cron job to fetch all messages from cache',
+  name: 'cron:messaging:gmail-partial-sync',
+  description:
+    'Starts a cron job to sync existing connected account messages and store them in the cache',
 })
-export class GmailFetchMessagesFromCacheCronCommand extends CommandRunner {
+export class GmailPartialSyncCronCommand extends CommandRunner {
   constructor(
     @Inject(MessageQueue.cronQueue)
     private readonly messageQueueService: MessageQueueService,
@@ -20,12 +23,10 @@ export class GmailFetchMessagesFromCacheCronCommand extends CommandRunner {
 
   async run(): Promise<void> {
     await this.messageQueueService.addCron<undefined>(
-      GmailFetchMessagesFromCacheCronJob.name,
+      GmailPartialSyncCronJob.name,
       undefined,
       {
-        repeat: {
-          every: 5000,
-        },
+        repeat: { pattern: GMAIL_PARTIAL_SYNC_CRON_PATTERN },
       },
     );
   }
