+++ packages/twenty-server/src/modules/messaging/crons/commands/messaging-cron-commands.module.ts
@@ -1,14 +1,16 @@
 import { Module } from '@nestjs/common';
 
-import { GoogleCalendarSyncCommand } from 'src/modules/calendar/commands/google-calendar-sync.command';
 import { ObjectMetadataRepositoryModule } from 'src/engine/object-metadata-repository/object-metadata-repository.module';
 import { ConnectedAccountObjectMetadata } from 'src/modules/connected-account/standard-objects/connected-account.object-metadata';
-import { GoogleCalendarSyncCronCommand } from 'src/modules/calendar/commands/crons/google-calendar-sync.cron.command';
-
+import { GmailFetchMessagesFromCacheCronCommand } from 'src/modules/messaging/crons/commands/gmail-fetch-messages-from-cache.cron.command';
+import { GmailPartialSyncCronCommand } from 'src/modules/messaging/crons/commands/gmail-partial-sync.cron.command';
 @Module({
   imports: [
     ObjectMetadataRepositoryModule.forFeature([ConnectedAccountObjectMetadata]),
   ],
-  providers: [GoogleCalendarSyncCommand, GoogleCalendarSyncCronCommand],
+  providers: [
+    GmailPartialSyncCronCommand,
+    GmailFetchMessagesFromCacheCronCommand,
+  ],
 })
-export class WorkspaceCalendarSyncCommandsModule {}
+export class MessagingCronCommandsModule {}
